Add dead body color restoration
Upon death, the dead body now returns to its original color instead of staying the twitch purple. Might update this later to choose a random player as the killer and only turn THEM purple, so it looks like Twitch is killing you and not itself, but I like how it is now. If people want it changed later that's how it should be done.